+++ tests/test_ux_features.py
@@ -344,11 +344,11 @@ class TestCombinedFeatures:
 class TestVersionUpdated:
     """Test that version is correct"""
 
-    def test_version_is_3_4_7(self):
-        """Test that version is 3.4.7"""
+    def test_version_is_3_4_8(self):
+        """Test that version is 3.4.8"""
         from cja_auto_sdr.generator import __version__
 
-        assert __version__ == "3.4.7"
+        assert __version__ == "3.4.8"
 
 
 class TestFormatAutoDetection: